fix: Browser. Page padding (#851)
* Browser. Page padding

* Browser. Rename BrowserMenuAnimationDelegate to BrowserAnimationDelegate

* Browser. Update paddingPageTween

* Browser. Menu animation no opacity

* Browser. Page lis background

* Browser. Return opacity

* Browser. Fix progress indicator

* Browser. Fix scroll

+++ lib/feature/browser_v2/screens/main/delegates/progress_indicator_delegate.dart
@@ -3,7 +3,7 @@ import 'package:flutter/cupertino.dart';
 abstract interface class BrowserProgressIndicatorUi {
   Animation<double> get animation;
 
-  void onProgressChanged(double progressValue);
+  void onProgressChanged(int progressValue);
 }
 
 class BrowserProgressIndicatorDelegate implements BrowserProgressIndicatorUi {
@@ -24,7 +24,7 @@ class BrowserProgressIndicatorDelegate implements BrowserProgressIndicatorUi {
   }
 
   @override
-  void onProgressChanged(double progressValue) {
+  void onProgressChanged(int progressValue) {
     _progressController.animateTo(progressValue / 100);
   }
